metaxy-example mkdocs plugin (#226)
Add mkdocs plugin that generates documentation for examples based on the `.example.yaml`​ file. 

It implements a few mkdocs directives that can be inserted into any `.md`​ file. 

It can display the original source code of the example or the result of applying the included `.patch`​ files (in order) to demonstrate code changes.